+++ npbench/benchmarks/cavity_flow/cavity_flow_dace.py
@@ -14,9 +14,9 @@ nx, ny, nit = (dc.symbol(s, dc.int64) for s in ('nx', 'ny', 'nit'))
 
 
 @dc.program
-def build_up_b(b: dc.float64[ny, nx], rho: dc.float64, dt: dc.float64,
-               u: dc.float64[ny, nx], v: dc.float64[ny, nx], dx: dc.float64,
-               dy: dc.float64):
+def build_up_b(b: dc.float32[ny, nx], rho: dc.float32, dt: dc.float32,
+               u: dc.float32[ny, nx], v: dc.float32[ny, nx], dx: dc.float32,
+               dy: dc.float32):
 
     b[1:-1,
       1:-1] = (rho * (1 / dt * ((u[1:-1, 2:] - u[1:-1, 0:-2]) / (2 * dx) +
@@ -28,8 +28,8 @@ def build_up_b(b: dc.float64[ny, nx], rho: dc.float64, dt: dc.float64,
 
 
 @dc.program
-def pressure_poisson(p: dc.float64[ny, nx], dx: dc.float64, dy: dc.float64,
-                     b: dc.float64[ny, nx]):
+def pressure_poisson(p: dc.float32[ny, nx], dx: dc.float32, dy: dc.float32,
+                     b: dc.float32[ny, nx]):
     pn = np.empty_like(p)
     pn[:] = p.copy()
 
@@ -47,10 +47,10 @@ def pressure_poisson(p: dc.float64[ny, nx], dx: dc.float64, dy: dc.float64,
 
 
 @dc.program
-def cavity_flow(nt: dc.int64, nit: dc.int64, u: dc.float64[ny, nx],
-                v: dc.float64[ny, nx], dt: dc.float64, dx: dc.float64,
-                dy: dc.float64, p: dc.float64[ny, nx], rho: dc.float64,
-                nu: dc.float64):
+def cavity_flow(nt: dc.int64, nit: dc.int64, u: dc.float32[ny, nx],
+                v: dc.float32[ny, nx], dt: dc.float32, dx: dc.float32,
+                dy: dc.float32, p: dc.float32[ny, nx], rho: dc.float32,
+                nu: dc.float32):
     un = np.empty_like(u)
     vn = np.empty_like(v)
     b = np.zeros((ny, nx))
