+++ npbench/benchmarks/go_fast/go_fast_dace.py
@@ -7,7 +7,7 @@ N = dc.symbol('N', dtype=dc.int64)
 
 
 @dc.program
-def go_fast(a: dc.float64[N, N]):
+def go_fast(a: dc.float32[N, N]):
     trace = 0.0
     for i in range(N):
         trace += np.tanh(a[i, i])
