+++ npbench/benchmarks/mandelbrot1/mandelbrot1_dace.py
@@ -11,27 +11,27 @@ XN, YN, N = (dc.symbol(s, dtype=dc.int64) for s in ['XN', 'YN', 'N'])
 
 
 @dc.program
-def linspace(start: dc.float64, stop: dc.float64, X: dc.float64[N]):
+def linspace(start: dc.float32, stop: dc.float32, X: dc.float32[N]):
     dist = (stop - start) / (N - 1)
     for i in dc.map[0:N]:
         X[i] = start + i * dist
 
 
 @dc.program
-def mandelbrot(xmin: dc.float64, xmax: dc.float64, ymin: dc.float64,
-               ymax: dc.float64, maxiter: dc.int64, horizon: dc.float64):
+def mandelbrot(xmin: dc.float32, xmax: dc.float32, ymin: dc.float32,
+               ymax: dc.float32, maxiter: dc.int64, horizon: dc.float32):
     # Adapted from https://www.ibm.com/developerworks/community/blogs/jfp/...
     #              .../entry/How_To_Compute_Mandelbrodt_Set_Quickly?lang=en
-    X = np.ndarray((XN, ), dtype=np.float64)
-    Y = np.ndarray((YN, ), dtype=np.float64)
+    X = np.ndarray((XN, ), dtype=np.float32)
+    Y = np.ndarray((YN, ), dtype=np.float32)
     linspace(xmin, xmax, X)
     linspace(ymin, ymax, Y)
     # C = X + np.reshape(Y, (YN, 1)) * 1j
-    C = np.ndarray((YN, XN), dtype=np.complex128)
+    C = np.ndarray((YN, XN), dtype=np.complex64)
     for i, j in dc.map[0:YN, 0:XN]:
         C[i, j] = X[j] + Y[i] * 1j
     N = np.zeros(C.shape, dtype=np.int64)
-    Z = np.zeros(C.shape, dtype=np.complex128)
+    Z = np.zeros(C.shape, dtype=np.complex64)
     for n in range(maxiter):
         I = np.less(np.absolute(Z), horizon)
         N[I] = n
