+++ npbench/benchmarks/polybench/atax/atax_dace.py
@@ -5,6 +5,6 @@ M, N = (dc.symbol(s, dtype=dc.int64) for s in ('M', 'N'))
 
 
 @dc.program
-def kernel(A: dc.float64[M, N], x: dc.float64[N]):
+def kernel(A: dc.float32[M, N], x: dc.float32[N]):
 
     return (A @ x) @ A
