+++ npbench/benchmarks/polybench/lu/lu_dace.py
@@ -5,7 +5,7 @@ N = dc.symbol('N', dtype=dc.int64)
 
 
 @dc.program
-def kernel(A: dc.float64[N, N]):
+def kernel(A: dc.float32[N, N]):
 
     for i in range(N):
         for j in range(i):
